Backport PR #5103 on branch v0.18.x (Replace np.bool, np.int, np.float with bool, int, float) (#5108)

+++ skimage/measure/_polygon.py
@@ -146,8 +146,8 @@ def subdivide_polygon(coords, degree=2, preserve_ends=False):
 
     mask_even, mask_odd = _SUBDIVISION_MASKS[degree]
     # divide by total weight
-    mask_even = np.array(mask_even, np.float) / (2 ** degree)
-    mask_odd = np.array(mask_odd, np.float) / (2 ** degree)
+    mask_even = np.array(mask_even, float) / (2 ** degree)
+    mask_odd = np.array(mask_odd, float) / (2 ** degree)
 
     even = signal.convolve2d(coords.T, np.atleast_2d(mask_even), mode=method,
                              boundary='wrap')
